+++ functions/_core/types.ts
@@ -188,25 +188,6 @@ export interface AppVehicleDetail extends AppVehicleProperties {
             };
         }>;
     };
-    route_geojson?: AppRouteGeoJSON;
-}
-
-export interface AppRouteFeature {
-    type: 'Feature';
-    geometry: {
-        type: 'LineString' | 'Point';
-        coordinates: number[] | number[][] | [number, number][];
-    };
-    properties: {
-        route_color: string;
-        is_terminal?: boolean;
-        [key: string]: unknown;
-    };
-}
-
-export interface AppRouteGeoJSON {
-    type: 'FeatureCollection';
-    features: AppRouteFeature[];
 }
 
 export interface AppDeparture {